Fixed problem with int values in set_numeric_option

+++ src/scalarizr/services/postgresql.py
@@ -729,7 +729,7 @@ class BasePGConfig(object):
 		
 	def set_numeric_option(self, option, number):
 		try:
-			assert not number or float(number)
+			assert not number or int(number)
 			self.set(option, str(number))
 		except ValueError:
 			raise ValueError('%s must be a number (got %s instead)' % (option, number))
@@ -752,8 +752,8 @@ class BasePGConfig(object):
 	
 	def get_numeric_option(self, option):
 		value = self.get(option)
-		assert not value or float(value)
-		return float(value) if value else 0
+		assert not value or int(value)
+		return int(value) if value else 0
 	
 	def save(self):
 		if self.data:
